+++ tests/testthat/test-flatten.R
@@ -16,3 +16,11 @@ test_that("can remove missing values", {
   expect_equal(str_flatten(c("a", NA)), NA_character_)
   expect_equal(str_flatten(c("a", NA), na.rm = TRUE), "a")
 })
+
+test_that("str_flatten_oxford removes comma iif necessary", {
+  expect_equal(str_flatten_comma(letters[1:2], ", or "), "a or b")
+
+  expect_equal(str_flatten_comma(letters[1:3], ", or "), "a, b, or c")
+  expect_equal(str_flatten_comma(letters[1:3], " or "), "a, b or c")
+  expect_equal(str_flatten_comma(letters[1:3]), "a, b, c")
+})