+++ R/utils.R
@@ -22,3 +22,6 @@ check_lengths <- function(string, pattern, replacement = NULL, error_call = call
 no_boundary <- function(call = caller_env()) {
   cli::cli_abort("{.arg pattern} can't be a boundary.", call = call)
 }
+no_empty <- function(call = caller_env()) {
+  cli::cli_abort("{.arg pattern} can't be the empty string ({.code \"\"}).", call = call)
+}