+++ NEWS.md
@@ -26,7 +26,8 @@
   no longer requires a pattern so you can use it to display strings with 
   special characters. It now highlights whitespace characters apart from space
   since otherwise they are often confusing. It's also now vectorised over both
-  `string` and `pattern` (#407).
+  `string` and `pattern` (#407). It defaults to displaying all matches, making
+  `str_view_all()` redundant (and hence deprecated) (#455).
 
 * New `vignette("from-base")` by @sastoudt provides a comprehensive comparison
   between base R functions and their stringr equivalents. It's designed to