+++ NEWS.md
@@ -7,15 +7,21 @@
   close to the tidyverse rules. There are only two major changes: 
   
     *  Only vectors of length 1 are recycled: previously,
-       `str_detect(letters, c("x", "y"))` worked, but it now errors.
+       (e.g.) `str_detect(letters, c("x", "y"))` worked, but it now errors.
        
     *  `str_c()` ignores `NULLs`, rather than treating them as length 0 
         vectors.
         
-    Additionally, many more non-vectorised arguments now throw errors,
-    rather than warnings, if supplied a vector.
+    Additionally, many more arguments now throw errors, rather than warnings, 
+    if supplied the wrong type of input.
 
 * `regex()` and friends now generate class names with `stringr_` prefix (#384).
+
+* `str_detect()`, `str_starts()`, `str_ends()` and `str_subset()` now error
+  when used with either an empty string (`""`) or a `boundary()`. These 
+  operations didn't really make sense (`str_detect(x, "")` returned `TRUE`
+  for all non-empty strings) and made it easy to make mistakes when programming
+  with patterns.
 
 ## New features
 
@@ -73,6 +79,9 @@
 * Better error message if you supply a non-string pattern (#378).
 
 * A new data source for `sentences` has fixed many small errors.
+
+* `str_extract()` and `str_exctract_all()` now work correctly when `pattern`
+  is a `boundary()`.
 
 * `str_flatten()` gains a `last` argument that optionally override the
   final separator (#377). It gains a `na.rm` argument to remove missing